requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-dotenv
requests
pydantic
pytest
httpx
google-genai
pandas
numpy
scikit-learn
seaborn
matplotlib
joblib
notebook
ipykernel
imbalanced-learn